change from long to BigInteger for results on day07

+++ day07/Data.cs
@@ -1,13 +1,14 @@
 namespace day07;
 
 using System.Text.RegularExpressions;
+using System.Numerics;
 
 public class Data
 {
     private static string pattern = @"(\d+):\s+(\d+(?:\s+\d+)+)";
 
-    public List<long> Results = new List<long>();
-    public List<List<long>> Operands = new List<List<long>>();
+    public List<BigInteger> Results = new List<BigInteger>();
+    public List<List<BigInteger>> Operands = new List<List<BigInteger>>();
 
     public Data(string filename)
     {
@@ -19,12 +20,12 @@ public class Data
             {
                 string value1 = match.Groups[1].Value;
                 string value2 = match.Groups[2].Value;
-                long result = long.Parse(value1);
+                BigInteger result = BigInteger.Parse(value1);
 
                 Results.Add(result);
 
                 string[] operandStrings = value2.Split(' ');
-                List<long> operands = operandStrings.Select(x => long.Parse(x)).ToList();
+                List<BigInteger> operands = operandStrings.Select(x => BigInteger.Parse(x)).ToList();
                 Operands.Add(operands);
             }
             else
